#15: Add Pagination in project table
