+++ utils/request.ts
@@ -1,9 +1,9 @@
-export async function get(input: string, query?: Record<string, string>) {
+export async function get<T>(input: string, query?: Record<string, any>) {
   const response = await fetch(`https://gantt-be.onrender.com${input}?${new URLSearchParams(JSON.parse(JSON.stringify(query)))}`, {
     method: "GET",
     headers: {
       "Content-Type": "application/json"
     }
   });
-  return await response.json();
+  return (await response.json()) as { data: T };
 }
